Created Customer models - Edited Event models - Created Interaction modeles
