+++ SimpleCRM.Models/EventModel/EventCreate.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SimpleCRM.Models.Event
+namespace SimpleCRM.Models.EventModel
 {
     public class EventCreate
     {
