+++ SimpleCRM.Models/EventModel/EventEdit.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SimpleCRM.Models.Event
+namespace SimpleCRM.Models.EventModel
 {
     public class EventEdit
     {
