+++ SimpleCRM.Models/EventModel/EventListItem.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SimpleCRM.Models.EventModel
+{
+    public class EventListItem
+    {
+        public int EventId { get; set; }
+        public string EventName { get; set; }
+    }
+}
