+++ SimpleCRM.Models/InteractionModel/InteractionListItem.cs
@@ -1,0 +1,16 @@
+﻿using SimpleCRM.Data;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SimpleCRM.Models.InteractionModel
+{
+    public class InteractionListItem
+    {
+        public int InteractionId { get; set; }
+        public int CustomerId { get; set; }
+        public Customer Customer { get; set; }
+    }
+}
