+++ SimpleCRM.Models/SimpleCRM.Models.csproj
@@ -55,18 +55,22 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Customer\CustomerCreate.cs" />
+    <Compile Include="Customer\CustomerDetail.cs" />
+    <Compile Include="Customer\CustomerEdit.cs" />
     <Compile Include="Customer\CustomerListItem.cs" />
+    <Compile Include="Event\EventCreate.cs" />
+    <Compile Include="Event\EventDetail.cs" />
+    <Compile Include="Event\EventEdit.cs" />
+    <Compile Include="Event\EventListItem.cs" />
+    <Compile Include="Interaction\InteractionCreate.cs" />
+    <Compile Include="Organization\OrganizationCreate.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
     <None Include="packages.config" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Event\" />
-    <Folder Include="Interaction\" />
-    <Folder Include="Organization\" />
-  </ItemGroup>
+  <ItemGroup />
   <ItemGroup>
     <ProjectReference Include="..\SimpleCRM.Data\SimpleCRM.Data.csproj">
       <Project>{9BA99FE2-1F2B-489B-BDB7-A61AE9715260}</Project>
